feat: implement logging functionality with CRUD operations and integrate logs into user actions

+++ src/app/api/users/[uuid]/route.ts
@@ -1,4 +1,8 @@
+import { LogsService } from '@/backend-service/services/log.service';
 import { UsersService } from '@/backend-service/services/users.service';
+import { auth } from '@/config/auth';
+import { CURD } from '@/model/list.model';
+import { isEmpty } from 'lodash';
 import { NextResponse } from 'next/server';
 
 export const PUT = async (
@@ -6,10 +10,25 @@ export const PUT = async (
   { params }: { params: Promise<{ uuid: string }> },
 ) => {
   try {
+    const logsService = new LogsService();
+
     const userService = new UsersService();
     const { uuid } = await params;
+    const session = await auth();
+    if (isEmpty(session)) {
+      return NextResponse.json(
+        { message: 'Unauthorized', status: 401 },
+        { status: 401 },
+      );
+    }
     const data = await request.json();
     const result = await userService.updateUser({ ...data, uuid });
+
+    await logsService.saveLogs(
+      session.user.deptID,
+      `อัปเดตผู้ใช้ UUID ${uuid}`,
+      CURD.UPDATE,
+    );
     return NextResponse.json(result, { status: result.status });
   } catch (error) {
     console.error(error);
@@ -44,8 +63,21 @@ export const DELETE = async (
 ) => {
   try {
     const userService = new UsersService();
+    const logsService = new LogsService();
+    const session = await auth();
+    if (isEmpty(session)) {
+      return NextResponse.json(
+        { message: 'Unauthorized', status: 401 },
+        { status: 401 },
+      );
+    }
     const { uuid } = await params;
     const result = await userService.deleteUser(uuid);
+    await logsService.saveLogs(
+      session.user.deptID,
+      `ลบผู้ใช้ UUID ${uuid}`,
+      CURD.DELETE,
+    );
     return NextResponse.json(result, { status: result.status });
   } catch (error) {
     console.error(error);
